new version

+++ src/main/java/ru/mokb/mars/requirements/database/PIMJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.mokb.mars.requirements.database.model.PIM;
 
 @Repository
-public interface PIMJpaRepository extends JpaRepository<PIM, Integer> {
+public interface PIMJpaRepository extends JpaRepository<PIM, Double> {
 }
